refactored tests to minimize collection deletions. sync part all passing

+++ tests/conftest.py
@@ -2,18 +2,21 @@
 Test fixtures
 """
 import os
+import math
 
 import pytest
 import uuid
-from typing import Dict, Iterable, Optional, AsyncIterable
+from typing import AsyncIterable, Dict, Iterable, List, Optional, TypeVar
 
 import pytest_asyncio
-from dotenv import load_dotenv
+# from dotenv import load_dotenv
 
 from astrapy.defaults import DEFAULT_KEYSPACE_NAME
 from astrapy.db import AstraDB, AstraDBCollection, AsyncAstraDB, AsyncAstraDBCollection
 
-load_dotenv()
+T = TypeVar("T")
+
+# load_dotenv()
 
 ASTRA_DB_APPLICATION_TOKEN = os.environ.get("ASTRA_DB_APPLICATION_TOKEN")
 ASTRA_DB_API_ENDPOINT = os.environ.get("ASTRA_DB_API_ENDPOINT")
@@ -22,8 +25,7 @@ ASTRA_DB_KEYSPACE = os.environ.get("ASTRA_DB_KEYSPACE", DEFAULT_KEYSPACE_NAME)
 # fixed
 TEST_WRITABLE_VECTOR_COLLECTION = "writable_v_col"
 TEST_READONLY_VECTOR_COLLECTION = "readonly_v_col"
-TEST_DISPOSABLE_VECTOR_COLLECTION = "disposable_v_col"
-TEST_DISPOSABLE_EMPTY_NONVECTOR_COLLECTION = "disposable_empty_col"
+TEST_WRITABLE_NONVECTOR_COLLECTION = "writable_nonv_col"
 
 VECTOR_DOCUMENTS = [
     {
@@ -50,6 +52,17 @@ VECTOR_DOCUMENTS = [
 ]
 
 
+def _batch_iterable(iterable: Iterable[T], batch_size: int) -> Iterable[Iterable[T]]:
+    this_batch = []
+    for entry in iterable:
+        this_batch.append(entry)
+        if len(this_batch) == batch_size:
+            yield this_batch
+            this_batch = []
+    if this_batch:
+        yield this_batch
+
+
 @pytest.fixture(scope="session")
 def astra_db_credentials_kwargs() -> Dict[str, Optional[str]]:
     return {
@@ -68,22 +81,12 @@ def astra_invalid_db_credentials_kwargs() -> Dict[str, Optional[str]]:
     }
 
 
-@pytest.fixture(scope="module")
-def cliff_uuid() -> str:
-    return str(uuid.uuid4())
-
-
-@pytest.fixture(scope="module")
-def vv_uuid() -> str:
-    return str(uuid.uuid4())
-
-
-@pytest.fixture(scope="module")
+@pytest.fixture(scope="session")
 def db(astra_db_credentials_kwargs: Dict[str, Optional[str]]) -> AstraDB:
     return AstraDB(**astra_db_credentials_kwargs)
 
 
-@pytest.fixture
+@pytest.fixture(scope="session")
 async def async_db(
     astra_db_credentials_kwargs: Dict[str, Optional[str]]
 ) -> AsyncIterable[AsyncAstraDB]:
@@ -98,8 +101,24 @@ def invalid_db(
     return AstraDB(**astra_invalid_db_credentials_kwargs)
 
 
-@pytest.fixture(scope="module")
-def writable_vector_collection(db: AstraDB) -> Iterable[AstraDBCollection]:
+@pytest.fixture(scope="session")
+def readonly_v_collection(db: AstraDB) -> Iterable[AstraDBCollection]:
+    collection = db.create_collection(
+        TEST_READONLY_VECTOR_COLLECTION,
+        dimension=2,
+    )
+
+    collection.truncate()
+    collection.insert_many(VECTOR_DOCUMENTS)
+
+    yield collection
+
+    if int(os.getenv("TEST_SKIP_COLLECTION_DELETE", "0")) == 0:
+        db.delete_collection(TEST_READONLY_VECTOR_COLLECTION)
+
+
+@pytest.fixture(scope="session")
+def writable_v_collection(db: AstraDB) -> Iterable[AstraDBCollection]:
     """
     This is lasting for the whole test. Functions can write to it,
     no guarantee (i.e. each test should use a different ID...
@@ -109,41 +128,58 @@ def writable_vector_collection(db: AstraDB) -> Iterable[AstraDBCollection]:
         dimension=2,
     )
 
-    collection.insert_many(VECTOR_DOCUMENTS)
-
     yield collection
 
-    db.delete_collection(TEST_WRITABLE_VECTOR_COLLECTION)
+    if int(os.getenv("TEST_SKIP_COLLECTION_DELETE", "0")) == 0:
+        db.delete_collection(TEST_WRITABLE_VECTOR_COLLECTION)
 
 
-@pytest_asyncio.fixture
-async def async_writable_vector_collection(
-    async_db: AsyncAstraDB,
-) -> AsyncIterable[AsyncAstraDBCollection]:
+@pytest.fixture(scope="function")
+def empty_v_collection(
+    writable_v_collection: AstraDBCollection,
+) -> Iterable[AstraDBCollection]:
+    """available empty to each test function."""
+    writable_v_collection.truncate()
+    yield writable_v_collection
+
+
+@pytest.fixture(scope="function")
+def disposable_v_collection(
+    writable_v_collection: AstraDBCollection,
+) -> Iterable[AstraDBCollection]:
+    """available prepopulated to each test function."""
+    writable_v_collection.truncate()
+    writable_v_collection.insert_many(VECTOR_DOCUMENTS)
+    yield writable_v_collection
+
+
+@pytest.fixture(scope="session")
+def writable_nonv_collection(db: AstraDB) -> Iterable[AstraDBCollection]:
     """
     This is lasting for the whole test. Functions can write to it,
     no guarantee (i.e. each test should use a different ID...
     """
-    collection = await async_db.create_collection(
-        TEST_WRITABLE_VECTOR_COLLECTION,
-        dimension=2,
-    )
-
-    await collection.insert_many(VECTOR_DOCUMENTS)
+    collection = db.create_collection(TEST_WRITABLE_NONVECTOR_COLLECTION)
 
     yield collection
 
-    await async_db.delete_collection(TEST_WRITABLE_VECTOR_COLLECTION)
+    if int(os.getenv("TEST_SKIP_COLLECTION_DELETE", "0")) == 0:
+        db.delete_collection(TEST_WRITABLE_NONVECTOR_COLLECTION)
+
+
+@pytest.fixture(scope="function")
+def empty_nonv_collection(
+    writable_nonv_collection: AstraDBCollection,
+) -> Iterable[AstraDBCollection]:
+    """available empty to each test function."""
+    writable_nonv_collection.truncate()
+    yield writable_nonv_collection
 
 
 @pytest.fixture(scope="module")
-def invalid_writable_vector_collection(
+def invalid_writable_v_collection(
     invalid_db: AstraDB,
 ) -> Iterable[AstraDBCollection]:
-    """
-    This is lasting for the whole test. Functions can write to it,
-    no guarantee (i.e. each test should use a different ID...
-    """
     collection = invalid_db.collection(
         TEST_WRITABLE_VECTOR_COLLECTION,
     )
@@ -151,85 +187,136 @@ def invalid_writable_vector_collection(
     yield collection
 
 
-@pytest.fixture(scope="module")
-def readonly_vector_collection(db: AstraDB) -> Iterable[AstraDBCollection]:
-    collection = db.create_collection(
-        TEST_READONLY_VECTOR_COLLECTION,
-        dimension=2,
-    )
+@pytest.fixture(scope="function")
+def pagination_v_collection(empty_v_collection: AstraDBCollection) -> Iterable[AstraDBCollection]:
 
-    collection.insert_many(VECTOR_DOCUMENTS)
+    INSERT_BATCH_SIZE = 20  # max 20, fixed by API constraints
+    N = 200  # must be EVEN
 
-    yield collection
+    def _mk_vector(index: int, n_total_steps: int) -> List[float]:
+        angle = 2 * math.pi * index / n_total_steps
+        return [math.cos(angle), math.sin(angle)]
 
-    db.delete_collection(TEST_READONLY_VECTOR_COLLECTION)
+    inserted_ids: Set[str] = set()
+    for i_batch in _batch_iterable(range(N), INSERT_BATCH_SIZE):
+        batch_ids = empty_v_collection.insert_many(
+            documents=[
+                {"_id": str(i), "$vector": _mk_vector(i, N)} for i in i_batch
+            ]
+        )["status"]["insertedIds"]
+        inserted_ids = inserted_ids | set(batch_ids)
+    assert inserted_ids == {str(i) for i in range(N)}
+
+    yield empty_v_collection
 
 
-@pytest.fixture
-async def async_readonly_vector_collection(
+@pytest_asyncio.fixture(scope="function")
+async def async_readonly_v_collection(
     async_db: AsyncAstraDB,
 ) -> AsyncIterable[AsyncAstraDBCollection]:
+    """
+    This is lasting for the whole test. Functions can write to it,
+    no guarantee (i.e. each test should use a different ID...
+    """
     collection = await async_db.create_collection(
         TEST_READONLY_VECTOR_COLLECTION,
         dimension=2,
     )
 
+    await collection.truncate()
     await collection.insert_many(VECTOR_DOCUMENTS)
 
     yield collection
 
-    await async_db.delete_collection(TEST_READONLY_VECTOR_COLLECTION)
+    if int(os.getenv("TEST_SKIP_COLLECTION_DELETE", "0")) == 0:
+        await async_db.delete_collection(TEST_READONLY_VECTOR_COLLECTION)
 
 
-@pytest.fixture(scope="function")
-def disposable_empty_nonvector_collection(db: AstraDB) -> Iterable[AstraDBCollection]:
-    collection = db.create_collection(
-        TEST_DISPOSABLE_EMPTY_NONVECTOR_COLLECTION,
-    )
-
-    yield collection
-
-    db.delete_collection(TEST_DISPOSABLE_EMPTY_NONVECTOR_COLLECTION)
-
-
-@pytest.fixture(scope="function")
-async def async_disposable_empty_nonvector_collection(
-    db: AsyncAstraDB,
-) -> AsyncIterable[AsyncAstraDBCollection]:
-    collection = await db.create_collection(
-        TEST_DISPOSABLE_EMPTY_NONVECTOR_COLLECTION,
-    )
-
-    yield collection
-
-    await db.delete_collection(TEST_DISPOSABLE_EMPTY_NONVECTOR_COLLECTION)
-
-
-@pytest.fixture(scope="function")
-def disposable_vector_collection(db: AstraDB) -> Iterable[AstraDBCollection]:
-    collection = db.create_collection(
-        TEST_DISPOSABLE_VECTOR_COLLECTION,
-        dimension=2,
-    )
-
-    collection.insert_many(VECTOR_DOCUMENTS)
-
-    yield collection
-
-    db.delete_collection(TEST_DISPOSABLE_VECTOR_COLLECTION)
-
-
-@pytest.fixture
-async def async_disposable_vector_collection(
+@pytest_asyncio.fixture(scope="session")
+async def async_writable_v_collection(
     async_db: AsyncAstraDB,
 ) -> AsyncIterable[AsyncAstraDBCollection]:
+    """
+    This is lasting for the whole test. Functions can write to it,
+    no guarantee (i.e. each test should use a different ID...
+    """
     collection = await async_db.create_collection(
-        TEST_DISPOSABLE_VECTOR_COLLECTION,
+        TEST_WRITABLE_VECTOR_COLLECTION,
         dimension=2,
     )
 
-    await collection.insert_many(VECTOR_DOCUMENTS)
+    yield collection
+
+    if int(os.getenv("TEST_SKIP_COLLECTION_DELETE", "0")) == 0:
+        await async_db.delete_collection(TEST_WRITABLE_VECTOR_COLLECTION)
+
+
+@pytest_asyncio.fixture(scope="function")
+async def async_empty_v_collection(
+    async_writable_v_collection: AsyncAstraDBCollection,
+) -> AsyncIterable[AsyncAstraDBCollection]:
+    """available empty to each test function."""
+    await async_writable_v_collection.truncate()
+    yield async_writable_v_collection
+
+
+@pytest_asyncio.fixture(scope="function")
+async def async_disposable_v_collection(
+    async_writable_v_collection: AsyncAstraDBCollection,
+) -> AsyncIterable[AsyncAstraDBCollection]:
+    """available prepopulated to each test function."""
+    await async_writable_v_collection.truncate()
+    await async_writable_v_collection.insert_many(VECTOR_DOCUMENTS)
+    yield writable_v_collection
+
+
+@pytest_asyncio.fixture(scope="session")
+async def async_writable_nonv_collection(
+    async_db: AsyncAstraDB,
+) -> AsyncIterable[AsyncAstraDBCollection]:
+    """
+    This is lasting for the whole test. Functions can write to it,
+    no guarantee (i.e. each test should use a different ID...
+    """
+    collection = await async_db.create_collection(
+        TEST_WRITABLE_VECTOR_COLLECTION,
+        dimension=2,
+    )
 
     yield collection
 
-    await async_db.delete_collection(TEST_DISPOSABLE_VECTOR_COLLECTION)
+    if int(os.getenv("TEST_SKIP_COLLECTION_DELETE", "0")) == 0:
+        await async_db.delete_collection(TEST_WRITABLE_VECTOR_COLLECTION)
+
+
+@pytest_asyncio.fixture
+async def async_empty_nonv_collection(
+    async_writable_nonv_collection: AsyncAstraDBCollection,
+) -> AsyncIterable[AsyncAstraDBCollection]:
+    """available empty to each test function."""
+    await async_writable_nonv_collection.truncate()
+    yield async_writable_nonv_collection
+
+
+@pytest_asyncio.fixture(scope="function")
+async def async_pagination_v_collection(async_empty_v_collection: AsyncAstraDBCollection) -> AsyncIterable[AsyncAstraDBCollection]:
+
+    INSERT_BATCH_SIZE = 20  # max 20, fixed by API constraints
+    N = 200  # must be EVEN
+
+    def _mk_vector(index: int, n_total_steps: int) -> List[float]:
+        angle = 2 * math.pi * index / n_total_steps
+        return [math.cos(angle), math.sin(angle)]
+
+    inserted_ids: Set[str] = set()
+    for i_batch in _batch_iterable(range(N), INSERT_BATCH_SIZE):
+        insert_response = await async_empty_v_collection.insert_many(
+            documents=[
+                {"_id": str(i), "$vector": _mk_vector(i, N)} for i in i_batch
+            ]
+        )
+        batch_ids = insert_response["status"]["insertedIds"]
+        inserted_ids = inserted_ids | set(batch_ids)
+    assert inserted_ids == {str(i) for i in range(N)}
+
+    yield async_empty_v_collection
